% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{bff_flatten_to_shape}
\alias{bff_flatten_to_shape}
\title{Flatten a mesh to a particular boundary shape}
\usage{
bff_flatten_to_shape(mesh, boundary_shape, normalise = TRUE)
}
\arguments{
\item{normalise}{}
}
\value{

}
\description{
Flatten a mesh to a particular boundary shape
}
