% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbff-package.R
\docType{data}
\name{hemisphere}
\alias{hemisphere}
\title{Hemisphere 3d Model}
\format{
A \code{mesh3d} object.
}
\source{
\url{https://github.com/GeometryCollective/boundary-first-flattening/tree/master/input}
}
\usage{
hemisphere
}
\description{
Hemisphere 3d Model
}
\keyword{datasets}
