% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbff-package.R
\docType{data}
\name{punctured_torus}
\alias{punctured_torus}
\title{Punctured Torus 3d Model}
\format{
A \code{mesh3d} object.
}
\source{
\url{https://github.com/GeometryCollective/boundary-first-flattening/tree/master/input}
}
\usage{
punctured_torus
}
\description{
Punctured Torus 3d Model
}
\keyword{datasets}
