% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r_logo_png}
\alias{r_logo_png}
\title{Get the R logo png file path.}
\usage{
r_logo_png()
}
\value{
A file path to the R logo png
}
\description{
Get the R logo png file path.
}
\examples{
imager::plot(imager::load.image(r_logo_png()))
}
