% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbff-package.R
\docType{data}
\name{spothead}
\alias{spothead}
\title{Spothead 3d Model}
\format{
A \code{mesh3d} object.
}
\source{
\url{https://github.com/GeometryCollective/boundary-first-flattening/tree/master/input}
}
\usage{
spothead
}
\description{
Spothead 3d Model
}
\keyword{datasets}
