% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{bff_vis_metrics}
\alias{bff_vis_metrics}
\title{Visualise the result of flattening a 3d mesh.}
\usage{
bff_vis_metrics(
  x,
  metric = c("area distortion", "vertice density", "curvature", "mesh"),
  ...
)
}
\arguments{
\item{x}{A \code{bff_flattened} or \code{bff_sphered} object}

\item{metric}{Which metric to display. One of "area distortion",
"vertice density", "curvature", or "mesh"}
}
\value{
None.
}
\description{
Visualise the result of flattening a 3d mesh.
}
\examples{
data(face)
## don't run the next line if you want to see the visualisation
options(rgl.useNULL = TRUE)
face_flat <- bff_flatten(face)
bff_vis_metrics(face_flat)
}
